+++ src/components/usecases/realtime_transcription/RealtimeTranscription.tsx
@@ -1,3 +1,4 @@
+import React, { useRef, useState, useEffect } from "react";
 import { MicOff, SettingsVoice, ArrowForward } from "@mui/icons-material";
 import {
   FormControl,
@@ -10,8 +11,6 @@ import {
   Tooltip,
   Typography,
 } from "@mui/material";
-import { useEffect, useState } from "react";
-import useSpeechToTextContinuous from "../../../hooks/useSpeechToTextContinuous";
 import MySpeechConfig from "../../../models/MySpeechConfig";
 import {
   SpeechServiceLanguagesNames,
@@ -22,6 +21,19 @@ import {
 import CustomIconButton from "../../common/CustomIconButton";
 import { UseCaseTemplateChildProps } from "../UseCaseTemplate";
 import TranscriptionResults from "./TranscriptionResults";
+import { SpeechToTextContinuous } from "./../../../hooks/speech_to_text_continuous/SpeechToTextContinuous";
+import { useDebounce, useDebouncedValue, useDidUpdate } from "rooks";
+import useTooltip from "../../../hooks/useTooltip";
+
+const autoStopRecognitionTimeout = 3 * 1000; //millis
+
+const debounce = (fn: Function, ms = 300) => {
+  let timeoutId: ReturnType<typeof setTimeout>;
+  return function (this: any, ...args: any[]) {
+    clearTimeout(timeoutId);
+    timeoutId = setTimeout(() => fn.apply(this, args), ms);
+  };
+};
 
 interface RealtimeTranscriptionProps extends UseCaseTemplateChildProps {
   speechConfig: MySpeechConfig;
@@ -33,66 +45,119 @@ const RealtimeTranscription: React.FC<RealtimeTranscriptionProps> = ({
   isSynthesizing,
   setError,
 }) => {
-  const speechToTextContinuous = useSpeechToTextContinuous(
-    speechConfig,
-    SpeechServiceLocale.German_Switzerland
+  const speechToTextContinuous = useRef(
+    new SpeechToTextContinuous(speechConfig)
   );
 
-  const [hasResults, setHasResults] = useState(false);
+  const [isRecognizing, setIsRecognizing] = useState(false);
+
+  const [recognitionLanguage, setRecognitionLanguage] = useState(
+    SpeechServiceLocale.German_Switzerland
+  );
+  const [translationTargetLanguage, setTranslationTargetLanguage] = useState(
+    SpeechTranslationLanguage.English
+  );
+
+  const [recognizingText, setRecognizingText] = useState("");
+  const [translatingText, setTranslatingText] = useState("");
+
+  const [recognizedResults, setRecognizedResults] = useState<string[]>([]);
+  const [translatedResults, setTranslatedResults] = useState<string[]>([]);
+
+  const stopRecognitionBecauseTimeoutToolTip = useTooltip(
+    `Aufnahme gestoppt, wegen Inaktivität für ${
+      autoStopRecognitionTimeout / 1000
+    } Sek`
+  );
+
+  const handleStartRecognition = () => {
+    speechToTextContinuous.current.startRecognition(
+      recognitionLanguage,
+      translationTargetLanguage
+    );
+    setError("");
+    setIsRecognizing(true);
+    resetTimeout.current();
+  };
+
+  const handleStopRecognition = () => {
+    speechToTextContinuous.current.stopRecognition();
+    setError("");
+    setIsRecognizing(false);
+  };
+
+  const handleTimeout = () => {
+    if (!isRecognizing) return;
+    handleStopRecognition();
+    stopRecognitionBecauseTimeoutToolTip.handleOpen();
+  };
+
+  const resetTimeout = useRef(
+    debounce(handleTimeout, autoStopRecognitionTimeout)
+  );
+
+  const hasResults = recognizedResults.length + translatedResults.length > 0;
 
   const handleSpeechRecognitionLanguageChange = (event: SelectChangeEvent) => {
-    speechToTextContinuous.setRecognitionLanguage(
-      event.target.value as SpeechServiceLocale
-    );
+    setRecognitionLanguage(event.target.value as SpeechServiceLocale);
   };
 
   const handleTranslationLanguageChange = (event: SelectChangeEvent) => {
-    speechToTextContinuous.setTranslationTargetLanguage(
+    setTranslationTargetLanguage(
       event.target.value as SpeechTranslationLanguage
     );
   };
 
-  // eslint-disable-next-line react-hooks/exhaustive-deps
   useEffect(() => {
     return () => {
-      speechToTextContinuous.sttFromMicStop();
+      speechToTextContinuous.current.closeRecognizer();
     };
   }, []);
 
-  useEffect(() => {
-    if (speechToTextContinuous.error.length > 0)
-      setError("Error in speech recognition: " + speechToTextContinuous.error);
-    else setError("");
-  }, [speechToTextContinuous.error, setError]);
+  const clearResults = () => {
+    setRecognizedResults([]);
+    setTranslatedResults([]);
+  };
+
+  speechToTextContinuous.current.error = (error, reason) => {
+    setIsRecognizing(false);
+    setError("Error in realtime transcription: " + error);
+  };
+
+  speechToTextContinuous.current.recognizing = (recognizing, translating) => {
+    setRecognizingText(recognizing);
+    setTranslatingText(translating);
+    resetTimeout.current();
+  };
+
+  speechToTextContinuous.current.recognized = (recognized, translated) => {
+    setRecognizingText("");
+    setTranslatingText("");
+    if (recognized && translated && recognized.length + translated.length > 0) {
+      setRecognizedResults([...recognizedResults, recognized]);
+      setTranslatedResults([...translatedResults, translated]);
+    }
+  };
 
   const startStopRecordingButton = () => (
     <CustomIconButton
       icon={
         <Tooltip
-          title={
-            speechToTextContinuous.stopRecognitionBecauseTimeoutToolTip.text
-          }
-          open={
-            speechToTextContinuous.stopRecognitionBecauseTimeoutToolTip.open
-          }
-          onClose={
-            speechToTextContinuous.stopRecognitionBecauseTimeoutToolTip
-              .handleClose
-          }
+          title={stopRecognitionBecauseTimeoutToolTip.text}
+          open={stopRecognitionBecauseTimeoutToolTip.open}
+          onClose={stopRecognitionBecauseTimeoutToolTip.handleClose}
         >
           <IconButton
             color="primary"
             size="large"
             onClick={
-              speechToTextContinuous.isRecognizing
-                ? speechToTextContinuous.sttFromMicStop
-                : speechToTextContinuous.sttFromMic
+              isRecognizing ? handleStopRecognition : handleStartRecognition
             }
             disabled={isSynthesizing}
             aria-label="Speak output"
             style={{ marginTop: "20px" }}
           >
-            {speechToTextContinuous.isRecognizing ? (
+            {isRecognizing ? (
               <MicOff fontSize="large" />
             ) : (
               <SettingsVoice fontSize="large" />
@@ -100,23 +165,15 @@ const RealtimeTranscription: React.FC<RealtimeTranscriptionProps> = ({
           </IconButton>
         </Tooltip>
       }
-      text={
-        speechToTextContinuous.isRecognizing
-          ? "Aufnahme stoppen"
-          : "Aufnahme starten"
-      }
+      text={isRecognizing ? "Aufnahme stoppen" : "Aufnahme starten"}
     />
   );
 
   const inputLanguageSelection = () => {
-    if (hasResults || speechToTextContinuous.isRecognizing)
+    if (hasResults || isRecognizing)
       return (
         <Typography variant="h6">
-          {
-            SpeechServiceLanguagesNames[
-            speechToTextContinuous.recognitionLanguage
-            ]
-          }
+          {SpeechServiceLanguagesNames[recognitionLanguage]}
         </Typography>
       );
     return (
@@ -128,7 +185,7 @@ const RealtimeTranscription: React.FC<RealtimeTranscriptionProps> = ({
           autoWidth
           labelId="language-input-select-label"
           id="language-input-select"
-          value={speechToTextContinuous.recognitionLanguage}
+          value={recognitionLanguage}
           label="Ausgabesprache"
           onChange={handleSpeechRecognitionLanguageChange}
         >
@@ -143,14 +200,10 @@ const RealtimeTranscription: React.FC<RealtimeTranscriptionProps> = ({
   };
 
   const outputLanguageSelection = () => {
-    if (hasResults || speechToTextContinuous.isRecognizing)
+    if (hasResults || isRecognizing)
       return (
         <Typography variant="h6" color="primary">
-          {
-            SpeechTranslationLanguagesNames[
-            speechToTextContinuous.translationTargetLanguage
-            ]
-          }
+          {SpeechTranslationLanguagesNames[translationTargetLanguage]}
         </Typography>
       );
     return (
@@ -162,7 +215,7 @@ const RealtimeTranscription: React.FC<RealtimeTranscriptionProps> = ({
           autoWidth
           labelId="language-select-label"
           id="language-select"
-          value={speechToTextContinuous.translationTargetLanguage}
+          value={translationTargetLanguage}
           label="Ausgabesprache"
           onChange={handleTranslationLanguageChange}
         >
@@ -178,16 +231,20 @@ const RealtimeTranscription: React.FC<RealtimeTranscriptionProps> = ({
 
   return (
     <div style={{ minHeight: "65vh" }}>
-      <Box sx={{ margin: '.5rem 0' }}>
-        {startStopRecordingButton()}
-      </Box>
-      <Box sx={{ margin: '1rem 0'}}>
-        <Box sx={{ display: 'flex', justifyContent: 'center', alignItems: 'center' }}>
+      <Box sx={{ margin: ".5rem 0" }}>{startStopRecordingButton()}</Box>
+      <Box sx={{ margin: "1rem 0" }}>
+        <Box
+          sx={{
+            display: "flex",
+            justifyContent: "center",
+            alignItems: "center",
+          }}
+        >
           {inputLanguageSelection()}
           <ArrowForward sx={{ width: 32, height: 32 }} />
           {outputLanguageSelection()}
         </Box>
-        {(hasResults || speechToTextContinuous.isRecognizing) && (
+        {(hasResults || isRecognizing) && (
           <Typography variant="body2" color="GrayText">
             Um die Sprachen zu wechseln, beende die Aufnahme und räume alle
             Aufnahmen auf.
@@ -198,8 +255,14 @@ const RealtimeTranscription: React.FC<RealtimeTranscriptionProps> = ({
       <TranscriptionResults
         synthesizeSpeech={synthesizeSpeech}
         isSynthesizing={isSynthesizing}
-        speechToTextContinuous={speechToTextContinuous}
-        handleHasResults={setHasResults}
+        recognitionLanguage={recognitionLanguage}
+        translationTargetLanguage={translationTargetLanguage}
+        recognizedResults={recognizedResults}
+        translatedResults={translatedResults}
+        recognizingText={recognizingText}
+        translatingText={translatingText}
+        clearResults={clearResults}
+        stopRecognition={handleStopRecognition}
       />
     </div>
   );